requirements.txt
APScheduler==3.11.1
asgiref @ file:///opt/miniconda3/conda-bld/asgiref_1760342057031/work
Django @ file:///opt/miniconda3/conda-bld/django_1760519133471/work
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
psutil==7.1.3
setuptools==80.9.0
sqlparse @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_444946bpsc/croot/sqlparse_1734704466504/work
tzlocal==5.3.1
wheel==0.45.1
